requirements.txt
requests>=2.28.0
asyncio>=3.4.3
typing>=3.7.4.3
argparse>=1.4.0
python-dotenv>=0.21.0 